refactor(auth): modularize authentication options and update API routes; remove unused middleware
