+++ apps/web/app/(app)/layout.tsx
@@ -43,7 +43,6 @@ export default function DashboardLayout({
           </Breadcrumb>
         </header>
         <main className="flex flex-1 flex-col gap-4 p-4">{children}</main>
-        <Toaster />
       </SidebarInset>
     </SidebarProvider>
   );
