+++ apps/web/app/login/page.tsx
@@ -1,10 +1,13 @@
 import { LoginForm } from "@/components/login-form";
+import { Suspense } from "react";
 
 export default function Page() {
   return (
     <div className="flex min-h-svh w-full items-center justify-center p-6 md:p-10">
       <div className="w-full max-w-sm">
-        <LoginForm />
+        <Suspense fallback={<div>Loading...</div>}>
+          <LoginForm />
+        </Suspense>
       </div>
     </div>
   );
